Correct the RMS-concat oracle and record SM121 qualification
Accumulate the BF16 reference projection in FP64 to preserve correctly rounded output ties across available architectures. Runtime kernels and exact top-1 checks remain unchanged. The five affected cases pass on SM120 and SM121; selected GB10 vocabulary, GEMV and loader tests pass, with original service containers restored.

Bind the frozen vocabulary source to 1,108 verified SM103 callables, retain all resource flags and register deltas, and record primary checkpoint identities and the separate V4.1 integration gap. Physical SM103 and complete model serving remain unqualified.

+++ b12x/sequence/mtp_feedback/reference.py
@@ -145,7 +145,9 @@ def rms_concat(
         ),
         dim=-1,
     )
-    return F.linear(joined, combined_fc_weight)
+    # BF16 GEMM may reduce partial sums in BF16 on some devices. The oracle
+    # uses FP64 accumulation so those reductions cannot change a BF16 tie.
+    return F.linear(joined.double(), combined_fc_weight.double()).to(joined.dtype)
 
 
 __all__.append("rms_concat")
